@@ -6,8 +6,8 @@
     "emailjs-com": "^3.2.0",
     "firebase": "^9.6.9",
     "npm": "^8.5.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-icons": "^4.3.1",
     "react-pdf": "^5.7.1",
     "react-router-dom": "^6.2.2"
