@@ -11,7 +11,7 @@ const postData = [
       "Okay, let's get back to web development. This is a field I've always wanted to get my hands (head) in deep and not in the dirty way! Yes, I've become a recent keyboard warrior coding away my evenings and nights learning as much as I can. Well, this blog in short is to summarize my experience as I take on this journey to envelop the web technologies and how they work and how I can use what I'm interested in to learn more and do more with techonology.",
       "To start off, I've been getting back into React, which is what this website was created with and some other cool packages. My goal is to improve my domain knowledge here, make this website better, and eventually, make a website for my loving significant other! Till I write again, I'll give you an update on what I've learned.",
     ],
-    "isShown": true,
+    "isShown": false,
   },
   {
     "id": "2",
@@ -24,6 +24,19 @@ const postData = [
       "I guess this brings it to the imposter syndrome people keep talking about for developers. Am I an imposter? Is this a self-report? From what I can tell looking at all the cool projects I've started, and mostly finished, I think I'm more of a crewmate these days than an imposter. Then when it comes to a new project, I get my hands on the keyboard again and start coding away and BAM, it hits me like an asteroid from outerspace, I don't really know what I'm truly coding. It always just feels like I've hit the tip of the iceberg...",
     ],
     "isShown": false,
+  },
+  {
+    "id": "3",
+    "title": "I think It's Going Great!",
+    "date": "April 2, 2022",
+    "body": [
+      "So far, it's been going great. I have a lot of mini projects. Each of them do their own little learning thing and they all still build on my foundation of web development. It has been stressful because there is still so much I do not know but at the same time its exciting to know that I have so much to learn!",
+      "The hard part for me I suppose now is in each of these mini projects, they are just 'mini' projects. Sooner or later, I'd like to be more proficient and actually do large projects and tie them all together. I have lots of cool ideas and the Internet is great for getting new inspriations. Learning how to fetch and pull data from public APIs has gotten easier, and understanding how it goes from one place to the next is something I'm starting to see more clearly.",
+      "Although I have a lot of projects, there are 2 in particular that I want to get really good at. One is for my beloved significant other and her at home hobby. This app is designed to be an e-commerce platform that shows what the products are, who she is, and how she does it. It's a big project and I expect to learn a lot from it. So far, the frontend is getting close to a finish but the data is lacking, and the custom CSS is also lacking. I have done a lot on it, but I just need the time to work on it.",
+      "The other project I have is this cool idea called Broken English. Basically, the app is going to be a peer to peer conversation. The first peer, the sender, will likely text in full English, and the other peer, the receiver, will see the text but changed to his/her tolerance of their own translation (ie. Spanish). The tolerance that the receiver sets is how much of the text is translated to another language. If you've ever heard 2nd generation people talk to their parents, it's sort of this half-a&*($#@d conversation where they are trying their best to speak in their parents' native language. This would be the same idea except its through text. The benefit of setting the tolerance (to start, would be like a beginner, intermediate, and finally advanced) is that everyone has their own pace to become fluent in other languages. This would be a full stack application where the front end appears to be like your normal Facebook Messenger chat app, or WhatsApp conversations, and the backend would take in the messages, translate them and send them out. I'm unsure of who is going to translate, but I would imagine it would be on the receiver side. The translation will be done through using FastAPI's Public Google Translate and hopefully I can learn a lot from it. The peer-to-peer framework will likely use Socket.IO which I've read great things about it. Finally, the frontend is going to be a React environment and the backend is a Node/Express Server.",
+      "With all this in mind, boy do I have a lot to go on. I just hope I can get more time to do these or learn on the job and apply what I learn to these side projects. It's all exciting and I wish I had done it sooner rather than later. On a final note: since working on my resume the first time using React to now, damn, React has gotten incredibly easier to work with. I'm excited to read more on React 17 and the transition to React 18!"
+    ],
+    "isShown": true,
   },
 
 ]
